add testcases & remove unused code
